requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==0.7.16
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.2.post1
arrow==1.3.0
astroid==3.1.0
asttokens==2.4.1
asyncssh==2.18.0
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
backports.tarfile==1.1.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.4.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
catboost==1.2.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cookiecutter==2.6.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diff-match-patch==20230430
dill==0.3.8
distro==1.9.0
docstring-to-markdown==0.15
docutils==0.21.1
docx2txt==0.8
entrypoints==0.4
executing==2.0.1
extra-streamlit-components==0.1.71
faiss-cpu==1.9.0
fastjsonschema==2.19.1
filelock==3.16.1
flake8==7.1.1
fonttools==4.51.0
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gpt4all==2.8.2
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.7
imagesize==1.4.1
importlib_metadata==7.1.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.3
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.10
keyring==25.1.0
kiwisolver==1.4.5
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langsmith==0.1.139
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
numpydoc==1.7.0
openai==1.53.0
orjson==3.10.11
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.0
plotly==5.21.0
pluggy==1.4.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.8.1b0
pydocstyle==6.3.0
pyflakes==3.2.0
PyGithub==2.4.0
Pygments==2.17.2
PyJWT==2.8.0
pylint==3.1.0
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.1.2
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
pyuca==1.2
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.3
QtAwesome==1.3.1
qtconsole==5.6.1
QtPy==2.4.1
referencing==0.34.0
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rope==1.13.0
rpds-py==0.18.0
Rtree==1.2.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.0
selenium==4.20.0
sentence-transformers==3.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
spyder==6.0.2
spyder-kernels==3.0.1
SQLAlchemy==2.0.35
stack-data==0.6.3
streamlit==1.33.0
streamlit-aggrid==1.0.4.post3
streamlit-authenticator==0.1.5
superqt==0.6.7
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
textdistance==4.6.1
threadpoolctl==3.5.0
three-merge==0.1.1
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
torch==2.5.1
tornado==6.4
tqdm==4.66.6
traitlets==5.14.2
transformers==4.46.1
trio==0.25.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
whatthepatch==1.0.5
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
yapf==0.40.2
yarl==1.17.1
zipp==3.18.1
